% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lambdass.R
\name{lambdass}
\alias{as.formals}
\alias{f.}
\alias{f.r}
\alias{f.s}
\alias{is.formals}
\alias{lambdass}
\title{Syntax sugar of making an anonymous function}
\usage{
as.formals(x, value = list(quote(expr = )))

is.formals(y)

f.r(..., env_ = parent.frame())

f.s(...)

f.(...)
}
\arguments{
\item{x}{x is coerced to character (internally STRSXP).}

\item{value}{a default value of a formal parameter.}

\item{y}{y is expected to be pairlist object with completely named}

\item{...}{If one argument is passed, that is parsed as lambda's body.
If two or more argument is passed, the last part is parsed as lambda's body
and the rest are parsed as formal parameters for lambda.}

\item{env_}{an environment in which call object is evaluated.}
}
\description{
Syntax sugar of making an anonymous function
}
\examples{
f.(x, x + 1) # => function(x) x + 1
 f.(x, f.(y, x + y)) #=>
 Reduce(f.(x, y, x + y), 1:10) # => 55
 Filter(f.(x, x \%\% 2 == 0), 1:10) # => c(2L, 4L, 6L, 8L, 10L)
}

